% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{containment_test}
\alias{containment_test}
\title{Test if species observations are contained in pixels.}
\usage{
containment_test(x, y, resolution)
}
\arguments{
\item{x}{2-column matrix with coordinate pairs.}

\item{y}{Spatial precision of each coordinate pair (in meters).}

\item{resolution}{Numeric vector with targeted pixel resolutions (in degrees).}
}
\value{
A logical vector.
}
\description{
Test if species observations are contained in pixels.
}
\details{
{Tests whether species observations (provided through \emph{x})
are contained in a pixel when accounting for their position in that pixel
and their coordinate uncertainty (given by \emph{y}). A coordinate pair
is said to be contained' if the buffer drawn around it by the respective
coordinate uncertainty does not extend beyond the boundaries of the pixel.}
}
