% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coordinate_precision}
\alias{coordinate_precision}
\title{Estimates the precision of lat/lon coordinates}
\usage{
coordinate_precision(x, y, verbose = T)
}
\arguments{
\item{x}{2-column matrix with observation coordinates.}

\item{y}{Numeric vector with uncertainty of coordinates in \emph{x}.}

\item{verbose}{Logical argument on whether progress should be reported.}
}
\value{
A logical vector.
}
\description{
Estimates the precision of lat/lon coordinates
}
\details{
{Measures the precision (in meters) of lat/lon
coordinate pairs (given by \emph{x}) based on their decimal precision.
Existing coordinate precision values (given by \emph{y} are added.)}
}
