% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dirichlet_alpha}
\alias{dirichlet_alpha}
\title{Estimates, or maps, Dirichlet distribution parameters}
\usage{
dirichlet_alpha(x, y)
}
\arguments{
\item{x}{numeric vector or a multi-layered \emph{SpatRaster} object.}

\item{y}{numeric vector or a multi-layered \emph{SpatRaster} object.}
}
\value{
A numeric vector or a set of SpatRaster objects.
}
\description{
Estimates, or maps, Dirichlet distribution parameters
}
\details{
{Estimates distribution parameters for each pixel, or across
all values, based on the mean (\emph{x}) and variance (\emph{y} of
estimate empirical probabilities of ecosystem occurrences.}
}
