% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{earth_radius}
\alias{earth_radius}
\title{Estimate Earth's radius}
\usage{
earth_radius(x)
}
\arguments{
\item{x}{A \emph{numeric} vector with latitudes (in degrees).}
}
\value{
Earth radius estimates for each element in \emph{x} (in meters).
}
\description{
Estimate Earth's radius
}
\details{
The radius of the Earth varies with latitude. Equatorial radius is ~6,378 km,
while the polar radius is ~6,357 km. At a given latitude, given by \emph{x},
the radius of the Earth is given by
\eqn{\sqrt{\frac{(a^2*cos(Φ))^2 + (b^2 * sin(Φ))^2}[a * cos(Φ))^2 + (b + sin(Φ))^2)}}
where \emph{a} is the equatorial radius, \emph{b} is the polar radius, and
\emph{Φ} is the latitude in radians.
}
