% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{entropy_threshold}
\alias{entropy_threshold}
\title{Find optimal entropy threshold}
\usage{
entropy_threshold(x, n = 100)
}
\arguments{
\item{x}{A \emph{sparRaster} object.}

\item{n}{Number of bins to create.}
}
\value{
A numeric element.
}
\description{
Find optimal entropy threshold
}
\details{
{For a \emph{spatRaster} object with probabilities between 0 and 1,
the function uses the workflow proposed by Kapur et al. (1985) to identify
the optimal maximum threshold to filter-out noise. For each map of empirical
ecosystem occurrence probabilities, we iterate through the distribution of those
probabilities defined by \emph{n} bins. At each bin (b), we distinguish pixels
with values below and above b and, for each of those groups (g), compute the
entropy (H) as  \eqn{\sum_{n=1}^{\infty} \frac{p_i}[P_b] * \log{\frac{p_b(g)}{P_b(g)}}}.
Here, \eqn{p_b(g)} is the normalized histogram value of a given bin belonging to the
focal group, and \eqn{P_b(g)} is the total probability mass function of the group.
After all iterations, the threshold which maximizes the total H between groups
is used as a cut-off value. Values above the threshold are preserved, consisting
of spatially coherent, high-probability pixels.}
}
\references{
Kapur, J. N. … Wong, A. K. C. (1985).
A new method for gray-level picture thresholding using the entropy of the histogram.
Computer Vision, Graphics, and Image Processing, 29(3), 273–285.
https://doi.org/10.1016/0734-189X(85)90125-2
}
