% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_filter.R
\name{location_filter}
\alias{location_filter}
\title{Location filter}
\usage{
location_filter(
  x,
  y,
  e,
  species,
  iso2c,
  resolution,
  g_ranges,
  e_ranges,
  elevation,
  method = "range",
  checklist = NULL
)
}
\arguments{
\item{x}{A character vector of x coordinates of species observations.}

\item{y}{A character vector of y coordinates of species observations.}

\item{e}{numeric vector with precision of each coordinate pairs formed between \emph{x} and \emph{y} (in meters).}

\item{species}{character vector with names of observed species.}

\item{iso2c}{character vector with ISO-2 country codes for each observations.}

\item{resolution}{target spatial resolution (in degrees)}

\item{g_ranges}{A \emph{SpatRaster} object.}

\item{e_ranges}{A list with 2-element vectors that represent elevation ranges of species occurrences.}

\item{elevation}{\emph{SpatRaster} of Digital Elevation Model (DEM).}

\item{method}{One of "range" or "checklist".}

\item{checklist}{A data frame}
}
\value{
a list
}
\description{
Location filter
}
\details{
For each coordinate pair of \emph{x} and \emph{y}, several tests are applied. First, the
function \link[CoordinateCleaner]{clean_coordinates} to control for the
misplacement of coordinates (e.g., over the center of countries, cities,
institutions providing access to the observation data). Second, \emph{g_ranges}
and \emph{e_ranges} are used to control, respectively, for expert knowledge
on the geographic and elevation ranges of species occurrences. If \emph{method}
is set to 'range', the content of \emph{g_ranges} is interpreted as an object
composed by multiple binary range maps, one for each unique element in
\emph{species}. Layers should be labeled with the respective species name. If
\emph{method} is set to 'checklist', \emph{g_ranges} is assumed to be a factorial
object. Note that the class names in the raster must correspond to those
in \emph{checklist}, which must also be provided as a 2-column data.frame
specifying the 'region'(s) of occurrence for each 'species'. When applying
either method, rows in \emph{x} are filtered if they are not supported by
the respective species geographic range/region and elevation range. Third,
a check evaluates whether each species observation, when buffered by its
coordinate precision, is contained in a pixel with the specified spatial
\emph{resolution}. This check is conducted over a regular global grid with
a resolution in degrees. When testing for containment, the precision of each
coordinate pair is inferred form the number of decimal places (as done in
\href{Open Street Map}{https://wiki.openstreetmap.org/wiki/Precision_of_coordinates}).
The final output is a list composed of:
\describe{
\item{flag}{logical vector with test results (TRUE for valid observations)}
\item{reason}{character vector of reason for failing the test (one of
'location', 'elevation range', 'geographic range', or 'pixel containment')}
\item{precision}{numeric vector with an updated estimate of the precision
of each coordinate pair.}
\item{cell}{cell number overlapping the each coordinate pair based on a
global raster grid with the specified resolution.}}
}
