% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_prior.R
\name{range_prior}
\alias{range_prior}
\title{Map Dirichlet distribution parameters}
\usage{
range_prior(x, y, lambda, path, verbose = T)
}
\arguments{
\item{x}{character vector with paths to species range maps in a grid format.}

\item{y}{character vector with ecosystem types for each entry in \emph{y}.}

\item{lambda}{numeric vector with distance decay parameters.}

\item{path}{directory where to store outputs.}

\item{verbose}{Logical argument. Should progress be reported?}
}
\value{
A SpatRaster.
}
\description{
Map Dirichlet distribution parameters
}
\details{
{Estimates empirical probabilities for a set of each types (given by
\emph{y}). First, for each unique ecosystem type, the range maps of inhabiting
species (given by \emph{x}) are summed. The sum is weighted using a distance
decay function, where the strength of decay is controlled by the argument
\emph{lambda}. Second, the empirical probability of each ecosystem type is
given by the ratio between the ecosystem-specific weighted count and the sum
of all weighted counts. Third, for each raster cell, shape parameters for
the Dirichlet conjugate prior (alpha) are on the mean and standard deviation
of the empirical ecosystem occurrence probabilities derive for each lambda.
Note that intermediate outputs (i.e., weighted sums, probability maps) are
written into \emph{path}.}
}
