% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_prior.R
\name{region_prior}
\alias{region_prior}
\title{Derive regional Dirichlet distribution parameters}
\usage{
region_prior(x, y, lambda)
}
\arguments{
\item{x}{2-column data.frame with 'species' and occurrence 'region'.}

\item{y}{{3-column data.frame with 'species', their 'habitat' associations,
and the thematic 'group' of each 'habitat' type in the respective typology.}}

\item{lambda}{distance decay parameter.}
}
\value{
A data.frame.
}
\description{
Derive regional Dirichlet distribution parameters
}
\details{
{Estimates empirical probabilities for each ecosystem type
distinguished in \emph{y}. First, for each unique ecosystem type, the
number of species using it as a habitat is recorded. This count is
weighted with a distance-decay function, where the strength of decay
is controlled by \emph{lambda}. Second, the empirical probability of
each ecosystem type is given by the ratio between the ecosystem-specific
weighted count and the sum of all weighted counts. The function returns
a data.frame with the following columns:
\describe{
\item{ecosystem}{ecosystem type used as a habitat.}
\item{w}{weighted species count.}
\item{p}{emprirical probability of target 'ecosystem' occurring in the given 'region'.}
\item{group}{Thematic group of the given 'ecosystem' types.}
\item{region}{region within which empirical probabiltiies were estimated.}
}}
}
