% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{weighted_sum}
\alias{weighted_sum}
\title{Derived weighted sum of distance weights}
\usage{
weighted_sum(x, lambda)
}
\arguments{
\item{x}{numeric vector of counts.}

\item{lambda}{numeric vector of distance-decay parameters.}
}
\value{
a named vector with a mean and standard deviation.
}
\description{
Derived weighted sum of distance weights
}
\details{
{Estimates the mean and standard deviation of the sum of
\emph{x} after applying an exponential distance decay function. The
number of iterations is equal to the length of \emph{lambda}, which
control the strength of the decay with higher values in \emph{x}. The
function returns the mean and variance of the total weights for each lambda.}
}
